backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sonora-backend"
version = "0.1.0"
description = "FastAPI backend and worker runtime for Sonora."
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.34",
  "bcrypt>=4.2,<5",
  "celery[redis]>=5.4",
  "email-validator>=2.2",
  "fastapi>=0.115",
  "passlib[bcrypt]>=1.7",
  "psycopg[binary]>=3.2",
  "pydantic-settings>=2.6",
  "python-jose[cryptography]>=3.3",
  "python-multipart>=0.0.12",
  "redis>=5.2",
  "sqlalchemy>=2.0",
  "uvicorn[standard]>=0.32",
  "yt-dlp>=2025.1.15"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.8"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
